agents/Langchain/requirements.txt
langchain>=1.0.0
langchain-openai>=1.0.0
langchain-core>=1.0.0
langchain-mcp-adapters>=0.1.0
python-dotenv>=1.0.0

